QtQml/qv4mm.cpp: Fix type of debug helper lastAllocRequestedSlots
Fix MSVC 64 developer build:
memory\qv4mm.cpp(653): error C2220: warning treated as error - no 'object' file generated
memory\qv4mm.cpp(653): warning C4267: '=': conversion from 'size_t' to 'int', possible loss of data
memory\qv4mm.cpp(689): warning C4267: '=': conversion from 'size_t' to 'int', possible loss of data

+++ src/qml/memory/qv4mm.cpp
@@ -643,7 +643,7 @@ MemoryManager::MemoryManager(ExecutionEngine *engine)
 }
 
 #ifndef QT_NO_DEBUG
-static int lastAllocRequestedSlots = 0;
+static size_t lastAllocRequestedSlots = 0;
 #endif
 
 Heap::Base *MemoryManager::allocString(std::size_t unmanagedSize)
